+++ projects/ids-enterprise-ng/src/lib/contextual-action-panel/soho-contextual-action-panel.ref.ts
@@ -284,10 +284,10 @@ export class SohoContextualActionPanelRef<T> {
     });
 
     // Add listeners to control events
-    this.jQueryElement.on('close', ((event: any, isCancelled: boolean) => { this.onClose(event, isCancelled); }));
-    this.jQueryElement.on('open', ((event: any) => { this.onOpen(event); }));
-    this.contextualactionpanel.panel.on('afterclose', ((event: any) => { this.onAfterClose(event); }));
-    this.contextualactionpanel.panel.on('afteropen', ((event: any) => { this.onAfterOpen(event); }));
+    this.jQueryElement.on('close.contextualactionpanel', ((event: any, isCancelled: boolean) => { this.onClose(event, isCancelled); }));
+    this.jQueryElement.on('open.contextualactionpanel', ((event: any) => { this.onOpen(event); }));
+    this.contextualactionpanel.panel.on('afterclose.contextualactionpanel', ((event: any) => { this.onAfterClose(event); }));
+    this.contextualactionpanel.panel.on('afteropen.contextualactionpanel', ((event: any) => { this.onAfterOpen(event); }));
 
     return this;
   }
@@ -407,7 +407,7 @@ export class SohoContextualActionPanelRef<T> {
     });
 
     if (this.jQueryElement) {
-      this.jQueryElement.off();
+      this.jQueryElement.off('close.contextualactionpanel open.contextualactionpanel');
     }
 
     this.contextualactionpanel.destroy();