+++ docs/CHANGELOG.md
@@ -4,9 +4,10 @@
 
 ### 7.1.0 Fixes
 
+- `[ContextualActionPanel]` The contextual action panel turned off all body events messing up other components attached to body so namespaced these.  `TJM` ([Issue #811](https://github.com/infor-design/enterprise-ng/issues/811))
+- `[Locale]` The types for locale were not incorrect, added fixes and enhanced them with new types.  `TJM` ([I #756](https://github.com/infor-design/enterprise-ng/issues/756))
 - `[Datagrid]` Added new aggregator types.  `TJM` ([#3752](https://github.com/infor-design/enterprise/issues/3752))
 - `[Datepicker]` Added newly added callback to disabled dates.  `TJM` ([#3462](https://github.com/infor-design/enterprise/issues/3462))
-- `[Locale]` The types for locale where incorrect so fixed them and added some new ones.  `TJM` ([I #756](https://github.com/infor-design/enterprise-ng/issues/756))
 - `[Wizard]` Prevent page refresh on selecting wizard ticks.  `BTHH` ([Pull Request #797](https://github.com/infor-design/enterprise-ng/pull/797))
 - `[Utils]` Added isInViewPort utils to types.  `TJM` ([#3738](https://github.com/infor-design/enterprise/issues/3738))